refactor(pi-task-scheduler): extract createSchedulerTools factory for host integration
Move 6 LLM tool definitions from extension.ts into standalone tools.ts factory.
Hosts can now import createSchedulerTools(scheduler) to register tools directly
without going through extension auto-discovery.
