+++ packages/pi-task-scheduler/src/index.ts
@@ -889,5 +889,6 @@ function rruleDayToCron(value: string | undefined): string {
     .join(',');
 }
 
+export { createSchedulerTools } from './tools.js';
 export { default } from './extension.js';
 export { FileSchedulerLock, JsonScheduledTaskStore } from './stores.js';
